BXMD-191/BW-132/BW-134 Feat: Add endpoint to check if a path is a file or directory modified:   server.js
Add a new endpoint '/isFileOrDirPath' to the Express server in server.js. This endpoint accepts a query parameter 'filePath', which is the path to be checked. The server responds with a JSON object containing the type of the path ('file' or 'dir') based on the existence and type of the file or directory at the specified path.